+++ src/app/features/features.routes.ts
@@ -1,0 +1,24 @@
+import { Routes } from '@angular/router';
+import { NavigationRoute } from '@shared/constant/navigation-route.const';
+
+export const featuresRoutes: Routes = [
+  {
+    path: '',
+    redirectTo: NavigationRoute.FEATURE.DASHBOARD,
+    pathMatch: 'full'
+  },
+  {
+    path: NavigationRoute.FEATURE.DASHBOARD,
+    loadComponent: () =>
+      import('@features/dashboard/dashboard.component').then(
+        c => c.DashboardComponent
+      )
+  },
+  {
+    path: NavigationRoute.FEATURE.MANAGE_USER.BASE,
+    loadChildren: () =>
+      import('@features/manage-user/manage-user.routes').then(
+        c => c.manageUserRoutes
+      )
+  }
+];