+++ src/app/features/manage-user/page/create-page/create-page.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-create-page',
+  imports: [],
+  templateUrl: './create-page.component.html',
+  styleUrl: './create-page.component.scss'
+})
+export class CreatePageComponent {}