+++ src/app/features/manage-user/page/list-page/list-page.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-list-page',
+  imports: [],
+  templateUrl: './list-page.component.html',
+  styleUrl: './list-page.component.scss'
+})
+export class ListPageComponent {}