+++ src/app/features/manage-user/page/update-page/update-page.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-update-page',
+  imports: [],
+  templateUrl: './update-page.component.html',
+  styleUrl: './update-page.component.scss'
+})
+export class UpdatePageComponent {}