+++ src/app/features/manage-user/page/view-page/view-page.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-view-page',
+  imports: [],
+  templateUrl: './view-page.component.html',
+  styleUrl: './view-page.component.scss'
+})
+export class ViewPageComponent {}